backend/audio_reader_standalone/requirements.txt
PyPDF2==3.0.1
google-generativeai==0.3.2
cartesia>=0.1.0
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
soundfile==0.12.1
websockets==12.0
PyAudio==0.2.14

